requirements.txt
chess==1.11.2
flask==3.1.2
flask-socketio==5.3.6
nmcli==1.5.0
pyserial==3.5
keyboard==0.13.5
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
eventlet==0.40.4
